WIP - Initial rework

+++ sim/rogue/items.go
@@ -31,9 +31,9 @@ var Tier6 = core.NewItemSet(core.ItemSet{
 		4: func(agent core.Agent, setBonusAura *core.Aura) {
 			// Increases the damage dealt by your Backstab, Sinister Strike, Mutilate, and Hemorrhage abilities by 6%.
 			setBonusAura.AttachSpellMod(core.SpellModConfig{
-				Kind:       core.SpellMod_DamageDone_Flat,
-				ClassMask:  RogueSpellBackstab | RogueSpellSinisterStrike | RogueSpellMutilate | RogueSpellHemorrhage,
-				FloatValue: 6.0,
+				Kind:      core.SpellMod_DamageDone_Flat,
+				ClassMask: RogueSpellBackstab | RogueSpellSinisterStrike | RogueSpellMutilate | RogueSpellHemorrhage,
+				IntValue:  6,
 			})
 		},
 	},
